[upgrade-webpack] production config updates:
Removing chunk configurations. It appears the default chunk
configuration works much better for our app than these.

A "chunk" is a group of modules represented in a minified JS file.
Previously, we had fewer, bigger chunks; now we more smaller chunks.
Basically one per page.

Another noteworthy change we've made is there is no "common" chunk.
Meaning, we no longer have a single chunk made up of modules that are
re-used by many other modules. Previously, this chunk was 1.21MB and
it would result in a lot of unused code being loaded for a given chunk
that depended on it.